Fix: NSPredicate was not accepting arguments being passed
In the Instantiation of the NSFetchResultsController, the NSPredicate was not being passed the arguments due to being in form of Double. Corrected all form of double lat/long into Float. Also changed in VirtualTouristModel as well.

+++ PinAnnotation+CoreDataProperties.swift
@@ -17,8 +17,8 @@ extension PinAnnotation
         return NSFetchRequest<PinAnnotation>(entityName: "PinAnnotation");
     }
 
-    @NSManaged public var latitude: Double
-    @NSManaged public var longitude: Double
+    @NSManaged public var latitude: Float
+    @NSManaged public var longitude: Float
     @NSManaged public var photos: NSSet?
 
 }
